Consolidate constants
