--- src/applications/disability-benefits/constants.js
@@ -1,5 +1,0 @@
-export const EBEN_526_PATH =
-  'https://www.ebenefits.va.gov/ebenefits/about/feature?feature=disability-compensation';
-
-export const BDD_INFO_URL =
-  '/disability/how-to-file-claim/when-to-file/pre-discharge-claim/';
