+++ src/applications/disability-benefits/wizard/pages/unable-to-file-bdd-production.jsx
@@ -5,7 +5,7 @@ import AlertBox from '@department-of-veterans-affairs/formation-react/AlertBox';
 import { pageNames } from './pageList';
 import { isLoggedIn as isLoggedInSelector } from 'platform/user/selectors';
 import recordEvent from 'platform/monitoring/record-event';
-import { EBEN_526_PATH, BDD_INFO_URL } from '../../constants';
+import { EBEN_526_PATH, BDD_INFO_URL } from '../../all-claims/constants';
 
 const dateFormat = 'MMMM DD, YYYY';
 const ninetyDays = moment()
